requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=5.4.0
requests>=2.26.0

# Optimization
pulp>=2.5.0

# Clustering
scikit-learn>=0.24.0

# Optional: For visualization
matplotlib>=3.4.0
plotly>=5.3.0

# Optional: For database integration
psycopg2-binary>=2.9.0

# Optional: For web API
flask>=2.0.0
flask-socketio>=5.1.0
flask-cors>=3.0.10

# Optional: For advanced features
pandas>=1.3.0

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=5.4.0
requests>=2.26.0

# Optimization
pulp>=2.5.0

# Clustering
scikit-learn>=0.24.0

# Web server (NEW for webapp)
flask>=2.0.0
flask-socketio>=5.1.0
flask-cors>=3.0.10
python-socketio>=5.4.0

# Optional: For visualization
matplotlib>=3.4.0
plotly>=5.3.0

# Optional: For database integration
psycopg2-binary>=2.9.0

# Optional: For advanced features
pandas>=1.3.0